// server/enums/abilityToWorkValue.ts
enum AbilityToWorkValue {
  LIMITED_BY_OFFENCE = 'LIMITED_BY_OFFENCE',
  CARING_RESPONSIBILITIES = 'CARING_RESPONSIBILITIES',
  NEEDS_WORK_ADJUSTMENTS_DUE_TO_HEALTH = 'NEEDS_WORK_ADJUSTMENTS_DUE_TO_HEALTH',
  UNABLE_TO_WORK_DUE_TO_HEALTH = 'UNABLE_TO_WORK_DUE_TO_HEALTH',
  LACKS_CONFIDENCE_OR_MOTIVATION = 'LACKS_CONFIDENCE_OR_MOTIVATION',
  REFUSED_SUPPORT_WITH_NO_REASON = 'REFUSED_SUPPORT_WITH_NO_REASON',
  RETIRED = 'RETIRED',
  NO_RIGHT_TO_WORK = 'NO_RIGHT_TO_WORK',
  NOT_SURE = 'NOT_SURE',
  OTHER = 'OTHER',
  NONE = 'NONE',
}

export default AbilityToWorkValue