// server/enums/skillsValue.ts
enum SkillsValue {
  COMMUNICATION = 'COMMUNICATION',
  POSITIVE_ATTITUDE = 'POSITIVE_ATTITUDE',
  RESILIENCE = 'RESILIENCE',
  SELF_MANAGEMENT = 'SELF_MANAGEMENT',
  TEAMWORK = 'TEAMWORK',
  THINKING_AND_PROBLEM_SOLVING = 'THINKING_AND_PROBLEM_SOLVING',
  WILLINGNESS_TO_LEARN = 'WILLINGNESS_TO_LEARN',
  OTHER = 'OTHER',
  NONE = 'NONE',
}

export default SkillsValue