// server/enums/additionalTrainingValue.ts
enum AdditionalTrainingValue {
  CSCS_CARD = 'CSCS_CARD',
  FULL_UK_DRIVING_LICENCE = 'FULL_UK_DRIVING_LICENCE',
  FIRST_AID_CERTIFICATE = 'FIRST_AID_CERTIFICATE',
  FOOD_HYGIENE_CERTIFICATE = 'FOOD_HYGIENE_CERTIFICATE',
  HEALTH_AND_SAFETY = 'HEALTH_AND_SAFETY',
  HGV_LICENCE = 'HGV_LICENCE',
  MACHINERY_TICKETS = 'MACHINERY_TICKETS',
  MANUAL_HANDLING = 'MANUAL_HANDLING',
  TRADE_COURSE = 'TRADE_COURSE',
  OTHER = 'OTHER',
  NONE = 'NONE',
}

export default AdditionalTrainingValue