// server/enums/typeOfWorkExperienceValue.ts
enum TypeOfWorkExperienceValue {
  OUTDOOR = 'OUTDOOR',
  CLEANING_AND_MAINTENANCE = 'CLEANING_AND_MAINTENANCE',
  CONSTRUCTION = 'CONSTRUCTION',
  DRIVING = 'DRIVING',
  BEAUTY = 'BEAUTY',
  HOSPITALITY = 'HOSPITALITY',
  TECHNICAL = 'TECHNICAL',
  MANUFACTURING = 'MANUFACTURING',
  OFFICE = 'OFFICE',
  RETAIL = 'RETAIL',
  SPORTS = 'SPORTS',
  WAREHOUSING = 'WAREHOUSING',
  EDUCATION_TRAINING = 'EDUCATION_TRAINING',
  WASTE_MANAGEMENT = 'WASTE_MANAGEMENT',
  OTHER = 'OTHER',
}

export default TypeOfWorkExperienceValue